requirements.txt
asgiref==3.9.1
django==5.2.6
django-environ==0.12.0
djangorestframework==3.16.1
environ==1.0
sqlparse==0.5.3
